Refactor output_handlers to a WorkflowOutputHandler class
