+++ ingestion/tests/integration/great_expectations/test_great_expectation_integration.py
@@ -38,7 +38,6 @@ from metadata.utils.time_utils import (
     get_end_of_day_timestamp_mill,
 )
 from metadata.workflow.metadata import MetadataWorkflow
-from metadata.workflow.workflow_output_handler import print_status
 
 Base = declarative_base()
