+++ ingestion/tests/integration/orm_profiler/system/test_redshift_system_metrics.py
@@ -47,7 +47,6 @@ from metadata.utils.time_utils import (
 )
 from metadata.workflow.metadata import MetadataWorkflow
 from metadata.workflow.profiler import ProfilerWorkflow
-from metadata.workflow.workflow_output_handler import print_status
 
 TESTS_ROOT_DIR = pathlib.Path(__file__).parent.parent.parent.parent
 REDSHIFT_CONFIG_FILE = "cli_e2e/database/redshift/redshift.yaml"
