+++ ingestion/tests/integration/orm_profiler/test_datalake_profiler_e2e.py
@@ -41,7 +41,6 @@ from metadata.utils.time_utils import (
 )
 from metadata.workflow.metadata import MetadataWorkflow
 from metadata.workflow.profiler import ProfilerWorkflow
-from metadata.workflow.workflow_output_handler import print_status
 
 SERVICE_NAME = Path(__file__).stem
 REGION = "us-west-1"
