+++ ingestion/tests/integration/profiler/test_nosql_profiler.py
@@ -45,7 +45,6 @@ from metadata.utils.test_utils import accumulate_errors
 from metadata.utils.time_utils import get_end_of_day_timestamp_mill
 from metadata.workflow.metadata import MetadataWorkflow
 from metadata.workflow.profiler import ProfilerWorkflow
-from metadata.workflow.workflow_output_handler import print_status
 
 SERVICE_NAME = Path(__file__).stem
