+++ openmetadata-airflow-apis/openmetadata_managed_apis/workflows/ingestion/profiler.py
@@ -28,7 +28,6 @@ from metadata.generated.schema.metadataIngestion.workflow import (
     WorkflowConfig,
 )
 from metadata.workflow.profiler import ProfilerWorkflow
-from metadata.workflow.workflow_output_handler import print_status
 
 
 def profiler_workflow(workflow_config: OpenMetadataWorkflowConfig):
@@ -48,7 +47,7 @@ def profiler_workflow(workflow_config: OpenMetadataWorkflowConfig):
 
     workflow.execute()
     workflow.raise_from_status()
-    print_status(workflow)
+    workflow.print_status()
     workflow.stop()
