+++ openmetadata-airflow-apis/openmetadata_managed_apis/workflows/ingestion/test_suite.py
@@ -28,7 +28,6 @@ from metadata.generated.schema.metadataIngestion.workflow import (
     WorkflowConfig,
 )
 from metadata.workflow.data_quality import TestSuiteWorkflow
-from metadata.workflow.workflow_output_handler import print_status
 
 
 def test_suite_workflow(workflow_config: OpenMetadataWorkflowConfig):
@@ -48,7 +47,7 @@ def test_suite_workflow(workflow_config: OpenMetadataWorkflowConfig):
 
     workflow.execute()
     workflow.raise_from_status()
-    print_status(workflow)
+    workflow.print_status()
     workflow.stop()
