+++ openmetadata-airflow-apis/openmetadata_managed_apis/workflows/ingestion/usage.py
@@ -32,7 +32,6 @@ from metadata.generated.schema.metadataIngestion.workflow import (
     Stage,
 )
 from metadata.workflow.usage import UsageWorkflow
-from metadata.workflow.workflow_output_handler import print_status
 
 
 def usage_workflow(workflow_config: OpenMetadataWorkflowConfig):
@@ -52,7 +51,7 @@ def usage_workflow(workflow_config: OpenMetadataWorkflowConfig):
 
     workflow.execute()
     workflow.raise_from_status()
-    print_status(workflow)
+    workflow.print_status()
     workflow.stop()
