+++ openmetadata-docs/content/partials/v1.5/deployment/run-connectors-class.md
@@ -15,7 +15,7 @@ def run():
     workflow = MetadataWorkflow.create(workflow_config)
     workflow.execute()
     workflow.raise_from_status()
-    print_status(workflow)
+    workflow.print_status()
     workflow.stop()
 ```
