+++ openmetadata-docs/content/v1.5.x-SNAPSHOT/connectors/pipeline/airflow/gcs-composer.md
@@ -65,7 +65,7 @@ from airflow.utils.dates import days_ago
 
 from metadata.workflow.metadata import MetadataWorkflow
 
-from metadata.workflow.workflow_output_handler import print_status
+ 
 
 default_args = {
     "owner": "user_name",
@@ -108,7 +108,7 @@ def metadata_ingestion_workflow():
     workflow = MetadataWorkflow.create(workflow_config)
     workflow.execute()
     workflow.raise_from_status()
-    print_status(workflow)
+    workflow.print_status()
     workflow.stop()
