+++ openmetadata-docs/content/v1.5.x-SNAPSHOT/deployment/ingestion/external/gcs-composer.md
@@ -52,7 +52,7 @@ from airflow.utils.dates import days_ago
 
 from metadata.workflow.metadata import MetadataWorkflow
 
-from metadata.workflow.workflow_output_handler import print_status
+ 
 
 default_args = {
     "owner": "user_name",
@@ -73,7 +73,7 @@ def metadata_ingestion_workflow():
     workflow = MetadataWorkflow.create(workflow_config)
     workflow.execute()
     workflow.raise_from_status()
-    print_status(workflow)
+    workflow.print_status()
     workflow.stop()
