+++ openmetadata-docs/content/v1.5.x-SNAPSHOT/deployment/upgrade/versions/110-to-120.md
@@ -230,12 +230,12 @@ We have reorganized the structure of the `Workflow` classes, which requires upda
 The `Workflow` class that you import can then be called as follows:
 
 ```python
-from metadata.workflow.workflow_output_handler import print_status
+ 
 
 workflow = workflow_class.create(workflow_config)
 workflow.execute()
 workflow.raise_from_status()
-print_status(workflow)  # This method has been updated. Before it was `workflow.print_status()`
+workflow.print_status()  # This method has been updated. Before it was `workflow.print_status()`
 workflow.stop()
 ```
