+++ openmetadata-docs/content/v1.5.x-SNAPSHOT/how-to-guides/data-insights/airflow-sdk.md
@@ -70,7 +70,7 @@ import yaml
 from datetime import timedelta
 from airflow import DAG
 from metadata.workflow.data_insight import DataInsightWorkflow
-from metadata.workflow.workflow_output_handler import print_status
+ 
 
 try:
     from airflow.operators.python import PythonOperator
@@ -98,7 +98,7 @@ def metadata_ingestion_workflow():
     workflow = DataInsightWorkflow.create(workflow_config)
     workflow.execute()
     workflow.raise_from_status()
-    print_status(workflow)
+    workflow.print_status()
     workflow.stop()
 
 with DAG(
